feat(Section): Section list 관리를 위한 일급컬렉션 추가
